+++ spec/request_handler_spec.rb
@@ -116,6 +116,7 @@ class IntegrationTestRequestHandlerWithMultiparts < RequestHandler::Base
   options do
     multipart do
       meta do
+        required true
         schema(Dry::Validation.JSON do
           configure do
             option :query_id
@@ -135,6 +136,7 @@ class IntegrationTestRequestHandlerWithMultiparts < RequestHandler::Base
       end
 
       file do
+        allowed true
       end
     end
   end