+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - update rubocop and fix danger check
 - drop support for ruby 2.1 (dry-types does not support it anymore either which makes supporting it here pointless)
 - adapt fieldset validation to allow all fields in addition to a specific enum
+- throw different errors for each parser, all new errors inherit from ExternalArgumentError
 
 ## 0.11.0